+++ src/index.tsx
@@ -47,8 +47,8 @@ export type { MovablePointProps } from "./interaction/MovablePoint"
 export { MovablePointDisplay } from "./display/MovablePointDisplay"
 export type { MovablePointDisplayProps } from "./display/MovablePointDisplay"
 
-export { useMovable } from "./interaction/useMovementInteraction"
-export type { UseMovable, UseMovableArguments } from "./interaction/useMovementInteraction"
+export { useMovable } from "./interaction/useMovable"
+export type { UseMovable, UseMovableArguments } from "./interaction/useMovable"
 
 export { useMovablePoint } from "./interaction/useMovablePoint"
 export type {
